Level generation from Text Files, Physics updates
I changed the level class a little bit, added a very primitive update
function for all Objects, and added more support for Physics from Box2D.
Also some more rearranging of files, but it shouldn't affect anyone
else's parts. I probably won't be able to get much else done this
weekend as I'll be out of town.

+++ level/level.h
@@ -2,20 +2,18 @@
 #ifndef LEVEL_H
 #define LEVEL_H
 #include <fstream>
-#include <sstream>
 #include <string>
-#include <cstdint>
 #include "object/object.h"
 
 class Level
 {
 	private:
 		//Private static class variables
-		const static std::uint8_t maximum_Number_Of_Objects = 10;
+		const static std::uint16_t maximum_Number_Of_Objects = 10;
 
 		//Private member variables
 		std::ifstream level_File;
-		std::uint8_t number_Of_Objects_In_Level;
+		std::uint16_t number_Of_Objects_In_Level;
 		Object level_Objects_Array[Level::maximum_Number_Of_Objects];
 
 	public:
